devops/requirements.txt
betterproto[compiler]
grpcio-tools
requests~=2.27.1
PyYAML~=6.0
gql[all]~=3.0.0
python-dateutil~=2.8.2